Fixed issues with landscape generation
Fixed two issues with the landscape generation: (1) ensured polygons now output the supply or demand, not neutral which was an error in the previous code, (2) now have fixed the supply gradient and generate the demand gradient relative to the supply gradient as a function of the interspersion of demand with supply. This ensures that the whole range from segregated, random, and interspersed is now possible. Only segregated and random were possible with the previous code.